+++ PGDatePicker/NSDate+PGCategory.h
@@ -8,14 +8,24 @@
 #import <Foundation/Foundation.h>
 
 @interface NSDate (PGCategory)
-+ (NSDate *)setYear:(NSInteger)year;
-+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month;
-+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day;
-+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute;
-+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute second:(NSInteger)second;
-+ (NSDate *)setHour:(NSInteger)hour minute:(NSInteger)minute;
-+ (NSDate *)setMinute:(NSInteger)minute second:(NSInteger)second;
-+ (NSDate *)setHour:(NSInteger)hour minute:(NSInteger)minute second:(NSInteger)second;
-+ (NSDate *)setMonth:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute;
-- (NSInteger)howManyDaysWithMonth;
+//+ (NSDate *)setYear:(NSInteger)year;
+//+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month;
+//+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day;
+//+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute;
+//+ (NSDate *)setYear:(NSInteger)year month:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute second:(NSInteger)second;
+//+ (NSDate *)setHour:(NSInteger)hour minute:(NSInteger)minute;
+//+ (NSDate *)setMinute:(NSInteger)minute second:(NSInteger)second;
+//+ (NSDate *)setHour:(NSInteger)hour minute:(NSInteger)minute second:(NSInteger)second;
+//+ (NSDate *)setMonth:(NSInteger)month day:(NSInteger)day hour:(NSInteger)hour minute:(NSInteger)minute;
+//- (NSInteger)howManyDaysWithMonth;
+
+/**
+ 根据dateComponents返回Date
+ @param dateComponents <#dateComponents description#>
+ @return <#return value description#>
+ */
++ (NSDate *)dateFromComponents:(NSDateComponents*)dateComponents;
+
+
+
 @end
