+++ PGDatePicker/PGDatePicker+MinuteAndSecond.m
@@ -54,7 +54,7 @@
 }
 
 - (void)minuteAndSecond_didSelectWithComponent:(NSInteger)component {
-    NSDateComponents *dateComponents = [self.calendar components:self.unitFlags fromDate:[NSDate date]];
+    NSDateComponents *dateComponents = [self.calendar components:unitFlags fromDate:[NSDate date]];
     NSString *str = [[self.pickerView textOfSelectedRowInComponent:0] componentsSeparatedByString:self.minuteString].firstObject;
     dateComponents.minute = [str integerValue];
     if (component == 0) {
