+++ PGDatePicker/PGDatePicker+MonthDay.m
@@ -43,7 +43,7 @@
 }
 
 - (void)monthDay_didSelectWithComponent:(NSInteger)component {
-    NSDateComponents *dateComponents = [self.calendar components:self.unitFlags fromDate:[NSDate date]];
+    NSDateComponents *dateComponents = [self.calendar components:unitFlags fromDate:[NSDate date]];
     NSString *str = [[self.pickerView textOfSelectedRowInComponent:0] componentsSeparatedByString:self.monthString].firstObject;
     dateComponents.month = [str integerValue];
     if (component == 0) {
