+++ PGDatePicker/PGDatePicker+MonthDayHour.m
@@ -64,7 +64,7 @@
 }
 
 - (void)monthDayHour_didSelectWithComponent:(NSInteger)component {
-    NSDateComponents *dateComponents = [self.calendar components:self.unitFlags fromDate:[NSDate date]];
+    NSDateComponents *dateComponents = [self.calendar components:unitFlags fromDate:[NSDate date]];
     NSString *str = [[self.pickerView textOfSelectedRowInComponent:0] componentsSeparatedByString:self.monthString].firstObject;
     dateComponents.month = [str integerValue];
     if (component == 0) {
