+++ PGDatePicker/PGDatePicker+MonthDayHourMinute.m
@@ -28,6 +28,7 @@
     NSString *minuteString = [self.pickerView textOfSelectedRowInComponent:3];
     minuteString = [minuteString componentsSeparatedByString:self.minuteString].firstObject;
     self.selectedComponents.minute = [minuteString integerValue];
+    
 }
 
 - (void)monthDayHourMinute_setDateWithComponents:(NSDateComponents *)components animated:(BOOL)animated {
@@ -80,7 +81,7 @@
 }
 
 - (void)monthDayHourMinute_didSelectWithComponent:(NSInteger)component {
-    NSDateComponents *dateComponents = [self.calendar components:self.unitFlags fromDate:[NSDate date]];
+    NSDateComponents *dateComponents = [self.calendar components:unitFlags fromDate:[NSDate date]];
     NSString *str = [[self.pickerView textOfSelectedRowInComponent:0] componentsSeparatedByString:self.monthString].firstObject;
     dateComponents.month = [str integerValue];
     if (component == 0) {
