+++ PGDatePicker/PGDatePicker+MonthDayHourMinuteSecond.m
@@ -96,7 +96,7 @@
 }
 
 - (void)monthDayHourMinuteSecond_didSelectWithComponent:(NSInteger)component {
-    NSDateComponents *dateComponents = [self.calendar components:self.unitFlags fromDate:[NSDate date]];
+    NSDateComponents *dateComponents = [self.calendar components:unitFlags fromDate:[NSDate date]];
     NSString *str = [[self.pickerView textOfSelectedRowInComponent:0] componentsSeparatedByString:self.monthString].firstObject;
     dateComponents.month = [str integerValue];
     if (component == 0) {
