+++ PGDatePicker/PGDatePicker+TimeAndSecond.m
@@ -70,7 +70,7 @@
 }
 
 - (void)timeAndSecond_didSelectWithComponent:(NSInteger)component {
-    NSDateComponents *dateComponents = [self.calendar components:self.unitFlags fromDate:[NSDate date]];
+    NSDateComponents *dateComponents = [self.calendar components:unitFlags fromDate:[NSDate date]];
     NSString *str = [[self.pickerView textOfSelectedRowInComponent:0] componentsSeparatedByString:self.hourString].firstObject;
     dateComponents.hour = [str integerValue];
     if (component == 0) {
